# api/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Bina Academy API",
    "description": "## نظرة عامة | Overview\n\nمجموعة الـ APIs الخاصة بمنصة **Bina Academy** لطلبة التوجيهي (الفرع العلمي والأدبي - الصف 11 و 12).\n\nAPI collection for **Bina Academy** tutoring platform (Tawjihi students - grades 11/12, scientific/literary branches).\n\n---\n\n### خصائص عامة | General Properties\n\n- جميع الردود JSON (`Content-Type: application/json`)\n- جميع أرقام الهواتف بالصيغة الفلسطينية `^05\\d{8}$` (مثال: `0591234567`)\n- التواريخ بصيغة ISO 8601\n- المصادقة بـ JWT Bearer (توكينات الطلاب 30 يوم، توكينات الإدمن 7 أيام)\n- الحد العام للطلبات: 100 طلب/دقيقة لكل IP (حسب `@nestjs/throttler`)\n\n### تدفّق العمل الموصى به | Recommended Mobile Flows\n\n1. **تسجيل طالب جديد**: POST `/auth/register` (الهاتف، كلمة المرور، الاسم الكامل، الصف، الفرع) → تصفح المواد مباشرة.\n2. **دخول طالب مسجل**: POST `/auth/login` (الهاتف وكلمة المرور) → تصفح المواد.\n3. **مشاهدة فيديو**: GET `/subjects/:id/videos` → GET `/videos/:id/stream` → POST `/videos/:id/mark-viewed`\n4. **إدمن**: POST `/auth/admin/login` → استعمل الـ token في كل endpoints الـ admin\n",
    "version": "1.0.0",
    "contact": {
      "name": "Bina Academy Backend Team"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Local development"
    },
    {
      "url": "http://10.0.2.2:3000/api",
      "description": "Android emulator (loopback to host)"
    },
    {
...
    }
  ],
...
  "paths": {
    "/auth/register": {
      "post": {
        "tags": ["Auth"],
        "summary": "تسجيل طالب جديد",
        "description": "يسجل طالب جديد في المنصة باستخدام رقم الجوال، الاسم، كلمة المرور، الصف، والفرع. يتم تلقائياً تفعيل باقة اشتراك تجريبية للوصول الأولي.\n\nRegisters a new Tawjihi student using phone, name, password, grade, and branch. Creates a TRIAL subscription automatically for early access.",
        "operationId": "registerStudent",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RegisterDto" },
              "examples": {
                "valid": {
                  "summary": "بيانات تسجيل طالب جديد",
                  "value": {
                    "phone": "0599222222",
                    "password": "student_password",
                    "name": "باسم سعيد",
                    "grade": "GRADE_11",
                    "branch": "SCIENTIFIC"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "تم تسجيل الحساب بنجاح ويحتوي على توكين JWT وبيانات الطالب",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AuthResponse" },
                "example": {
                  "token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
                  "user": {
                    "id": "cmrcd5l5z0000bgsuu0noi98n",
                    "phone": "0599222222",
                    "name": "باسم سعيد",
                    "grade": "GRADE_11",
                    "branch": "SCIENTIFIC",
                    "parentPhone": null,
                    "role": "STUDENT",
                    "createdAt": "2026-07-08T17:40:23.015Z"
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "تسجيل دخول طالب",
        "description": "يسجل دخول الطالب باستخدام رقم الجوال وكلمة المرور المشفرة مسبقاً.\n\nAuthenticates a student using phone number and password.",
        "operationId": "loginStudent",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginDto" },
              "examples": {
                "valid": {
                  "summary": "تسجيل دخول بالرقم وكلمة المرور",
                  "value": { "phone": "0599222222", "password": "student_password" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "دخول ناجح — يحتوي على توكين JWT للطالب",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AuthResponse" },
                "example": {
                  "token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
                  "user": {
                    "id": "cmrcd5l5z0000bgsuu0noi98n",
                    "phone": "0599222222",
                    "name": "باسم سعيد",
                    "grade": "GRADE_11",
                    "branch": "SCIENTIFIC",
                    "parentPhone": null,
                    "role": "STUDENT",
                    "createdAt": "2026-07-08T17:40:23.015Z"
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": {
            "description": "رقم الهاتف أو كلمة المرور غير صحيحة",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": {
                  "statusCode": 401,
                  "message": "رقم الهاتف أو كلمة المرور غير صحيحة"
                }
              }
            }
          }
        }
      }
    },
    "/auth/setup-profile": {
      "post": {
        "tags": ["Auth"],
        "summary": "إكمال ملف الطالب (الصف + الفرع)",
        "description": "يضبط صف وفرع الطالب. إذا لم يكن لديه اشتراك فعّال، يتم إنشاء اشتراك **TRIAL** تلقائياً (إن وُجدت خطة TRIAL في DB). يتطلب توكين STUDENT.\n\nSets the student's grade and branch. If no active subscription exists, a TRIAL subscription is auto-created (if the TRIAL plan exists in DB). Requires STUDENT JWT.",
        "operationId": "setupProfile",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SetupProfileDto" },
              "examples": {
                "scientific": {
                  "summary": "طالب علمي صف ثاني عشر",
                  "value": { "grade": "GRADE_12", "branch": "SCIENTIFIC" }
                },
                "literary": {
                  "summary": "طالب أدبي صف حادي عشر",
                  "value": { "grade": "GRADE_11", "branch": "LITERARY" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "تم تحديث الملف وإنشاء اشتراك تجريبي (إن أمكن)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": { "$ref": "#/components/schemas/User" }
                  }
                },
                "example": {
                  "user": {
                    "id": "clv123abc",
                    "phone": "0591234567",
                    "name": null,
                    "grade": "GRADE_12",
                    "branch": "SCIENTIFIC",
                    "parentPhone": null,
                    "role": "STUDENT",
                    "createdAt": "2026-06-28T15:00:00.000Z"
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/auth/admin/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "تسجيل دخول الإدمن",
        "description": "يحقّق من بريد الإدمن وكلمة المرور ويعود بـ JWT ADMIN (صالح 7 أيام). **مختلف عن توكين الطالب**: يتطلب `ADMIN_JWT_SECRET` (أو `JWT_SECRET` كاحتياط).\n\nVerifies admin email and password, returns an ADMIN JWT (7-day expiry). **Different from student tokens**: uses `ADMIN_JWT_SECRET` (or `JWT_SECRET` as fallback).",
        "operationId": "adminLogin",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AdminLoginDto" },
              "example": {
                "email": "admin@bina-academy.ps",
                "password": "CHANGE_ME"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "تسجيل دخول ناجح",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AdminLoginResponse" },
                "example": {
                  "token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
                  "admin": {
                    "id": "admin001",
                    "email": "admin@bina-academy.ps",
                    "name": "مدير النظام"
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": {
            "description": "بيانات الدخول غير صحيحة",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": {
                  "statusCode": 401,
                  "message": "بيانات الدخول غير صحيحة"
                }
              }
            }
          }
        }
      }
    },
...
    "/cart": {
      "get": {
        "tags": ["Cart"],
        "summary": "عرض سلة التسوق | Get Cart",
        "description": "يعود بالمواد الموجودة في السلة ومجموع أسعارها. يتطلب توكين STUDENT.\n\nReturns items in the student's cart and the total price. Requires STUDENT JWT.",
        "operationId": "getCart",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "سلة التسوق بنجاح",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CartResponse" },
                "example": {
                  "items": [
                    {
                      "id": "cart_item_001",
                      "subjectId": "sub_math",
                      "name": "الرياضيات",
                      "grade": "GRADE_12",
                      "branch": "SCIENTIFIC",
                      "priceIls": 80.0,
                      "teacherName": "أ. خالد عمر"
                    }
                  ],
                  "totalPriceIls": 80.0
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/cart/add": {
      "post": {
        "tags": ["Cart"],
        "summary": "إضافة مادة للسلة | Add to Cart",
        "description": "يضيف مادة إلى سلة التسوق. يفحص وجود المادة، عدم اشتراك الطالب فيها باشتراك مدفوع فعال، وعدم وجودها في السلة مسبقاً. يتطلب توكين STUDENT.\n\nAdds a subject to the cart. Checks subject existence, active paid subscriptions, and previous inclusion. Requires STUDENT JWT.",
        "operationId": "addToCart",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AddCartItemDto" },
              "example": { "subjectId": "sub_math" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "تمت الإضافة بنجاح",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string", "example": "تم إضافة المادة للسلة بنجاح" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "المادة مضافة مسبقاً أو هناك اشتراك فعال",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "examples": {
                  "alreadySubscribed": {
                    "summary": "مشترك بالفعل",
                    "value": { "statusCode": 400, "message": "أنت مشترك بالفعل في هذه المادة واشتراكك فعال" }
                  },
                  "alreadyInCart": {
                    "summary": "في السلة بالفعل",
                    "value": { "statusCode": 400, "message": "المادة موجودة بالفعل في السلة" }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": {
            "description": "المادة غير موجودة",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": { "statusCode": 404, "message": "المادة غير موجودة" }
              }
            }
          }
        }
      }
    },
    "/cart/remove/{subjectId}": {
      "delete": {
        "tags": ["Cart"],
        "summary": "حذف مادة من السلة | Remove from Cart",
        "description": "يحذف مادة من سلة التسوق. يتطلب توكين STUDENT.\n\nRemoves a subject from the cart. Requires STUDENT JWT.",
        "operationId": "removeFromCart",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "subjectId",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "example": "sub_math" }
          }
        ],
        "responses": {
          "200": {
            "description": "تم الحذف بنجاح",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string", "example": "تم حذف المادة من السلة بنجاح" }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": {
            "description": "المادة غير موجودة في السلة",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": { "statusCode": 404, "message": "المادة غير موجودة في السلة" }
              }
            }
          }
        }
      }
    },
    "/cart/checkout": {
      "post": {
        "tags": ["Cart"],
        "summary": "إتمام الشراء والدفع الافتراضي | Checkout Cart",
        "description": "يشتري المواد الموجودة في السلة، يربطها باشتراك مدفوع فعال (أو ينشئ اشتراك شهري جديد)، يفرغ السلة، ويحسب المجموع. يتطلب توكين STUDENT.\n\nPerforms a checkout of the cart using simulated wallet payment. Activates subscription for these subjects and clears the cart. Requires STUDENT JWT.",
        "operationId": "checkoutCart",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "تم الشراء وتفعيل الاشتراك بنجاح",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CheckoutResponse" },
                "example": {
                  "message": "تم الشراء وتفعيل الاشتراك بنجاح عن طريق المحفظة الإلكترونية الافتراضية",
                  "totalPriceIls": 80.0,
                  "subscriptionId": "sub_001",
                  "endDate": "2026-07-28T15:00:00.000Z",
                  "subjects": ["الرياضيات"]
                }
              }
            }
          },
          "400": {
            "description": "السلة فارغة",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": { "statusCode": 400, "message": "السلة فارغة" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Bearer.\n\n- توكين STUDENT: يولَّد من `/auth/verify-otp` (صالح 30 يوم، يستخدم `JWT_SECRET`)\n- توكين ADMIN: يولَّد من `/auth/admin/login` (صالح 7 أيام، يستخدم `ADMIN_JWT_SECRET` أو `JWT_SECRET` كاحتياط)\n\n> ⚠️ **مهم**: لا تستخدم توكين STUDENT في endpoints الـ Admin والعكس صحيح — كلاهما سيرد بـ 401."
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Validation error - عادةً من class-validator على الحقول المطلوبة/الصيغة",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": {
              "statusCode": 400,
              "message": ["phone must match /^05\\d{8}$/ regular expression"],
              "error": "Bad Request"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "التوكن مفقود/منتهي/غير صالح أو الـ role لا يطابق الـ endpoint",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "examples": {
              "missingToken": { "summary": "بدون توكن", "value": { "statusCode": 401, "message": "Unauthorized" } },
              "badRole": { "summary": "توكين لا يطابق الـ role", "value": { "statusCode": 401, "message": "جلسة غير صالحة" } }
            }
          }
        }
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "statusCode": { "type": "integer", "example": 400 },
          "message": { "oneOf": [{ "type": "string" }, { "type": "array", "items": { "type": "string" } }] },
          "error": { "type": "string", "example": "Bad Request" }
        },
        "required": ["statusCode", "message"]
      },
...
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "example": "clv123abc" },
          "phone": { "type": "string", "example": "0591234567", "description": "رقم هاتف فلسطيني 10 أرقام يبدأ بـ 05" },
          "name": { "type": "string", "nullable": true, "example": "أحمد محمود" },
          "grade": { "type": "string", "nullable": true, "enum": ["GRADE_11", "GRADE_12"] },
          "branch": { "type": "string", "nullable": true, "enum": ["SCIENTIFIC", "LITERARY"] },
          "parentPhone": { "type": "string", "nullable": true, "example": "0599876543" },
          "role": { "type": "string", "enum": ["STUDENT", "TEACHER", "ADMIN"], "example": "STUDENT" },
          "createdAt": { "type": "string", "format": "date-time" }
        },
        "required": ["id", "phone", "role", "createdAt"]
      },
...
      "RegisterDto": {
        "type": "object",
        "required": ["phone", "password", "name", "grade", "branch"],
        "properties": {
          "phone": { "type": "string", "pattern": "^05\\d{8}$", "example": "0599222222", "description": "رقم هاتف فلسطيني - 10 أرقام يبدأ بـ 05" },
          "password": { "type": "string", "minLength": 6, "maxLength": 20, "example": "student_password", "description": "كلمة المرور الخاصة بالطالب" },
          "name": { "type": "string", "example": "باسم سعيد", "description": "الاسم الكامل للطالب" },
          "grade": { "type": "string", "enum": ["GRADE_11", "GRADE_12"], "example": "GRADE_11" },
          "branch": { "type": "string", "enum": ["SCIENTIFIC", "LITERARY"], "example": "SCIENTIFIC" },
          "parentPhone": { "type": "string", "pattern": "^05\\d{8}$", "example": "0599987654", "description": "رقم هاتف ولي الأمر (اختياري)" }
        }
      },
      "LoginDto": {
        "type": "object",
        "required": ["phone", "password"],
        "properties": {
          "phone": { "type": "string", "pattern": "^05\\d{8}$", "example": "0599222222" },
          "password": { "type": "string", "example": "student_password" }
        }
      },
      "SetupProfileDto": {
        "type": "object",
        "required": ["grade", "branch"],
        "properties": {
          "grade": { "type": "string", "enum": ["GRADE_11", "GRADE_12"] },
          "branch": { "type": "string", "enum": ["SCIENTIFIC", "LITERARY"] }
        }
      },
      "AdminLoginDto": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": { "type": "string", "format": "email" },
          "password": { "type": "string", "format": "password" }
        }
      },
...
      "AuthResponse": {
        "type": "object",
        "properties": {
          "token": { "type": "string", "description": "JWT - صالح 30 يوم" },
          "user": { "$ref": "#/components/schemas/User" }
        }
      },
      "AdminLoginResponse": {
        "type": "object",
        "properties": {
          "token": { "type": "string", "description": "ADMIN JWT - صالح 7 أيام" },
          "admin": {
            "type": "object",
            "properties": {
              "id": { "type": "string" },
              "email": { "type": "string", "format": "email" },
              "name": { "type": "string" }
            }
          }
        }
      },
...
      "AddCartItemDto": {
        "type": "object",
        "required": ["subjectId"],
        "properties": {
          "subjectId": { "type": "string", "example": "sub_math" }
        }
      },
      "CartItem": {
        "type": "object",
        "required": ["id", "subjectId", "name", "grade", "branch", "priceIls"],
        "properties": {
          "id": { "type": "string" },
          "subjectId": { "type": "string" },
          "name": { "type": "string" },
          "grade": { "type": "string", "enum": ["GRADE_11", "GRADE_12"] },
          "branch": { "type": "string", "enum": ["SCIENTIFIC", "LITERARY"] },
          "priceIls": { "type": "number" },
          "teacherName": { "type": "string", "nullable": true }
        }
      },
      "CartResponse": {
        "type": "object",
        "required": ["items", "totalPriceIls"],
        "properties": {
          "items": { "type": "array", "items": { "$ref": "#/components/schemas/CartItem" } },
          "totalPriceIls": { "type": "number" }
        }
      },
      "CheckoutResponse": {
        "type": "object",
        "required": ["message", "totalPriceIls", "subscriptionId", "endDate", "subjects"],
        "properties": {
          "message": { "type": "string" },
          "totalPriceIls": { "type": "number" },
          "subscriptionId": { "type": "string" },
          "endDate": { "type": "string", "format": "date-time" },
          "subjects": { "type": "array", "items": { "type": "string" } }
        }
      },
...
    }
  }
}